hotfix: 미흡한 validation 처리
