+++ api/src/main/java/shop/ink3/api/book/book/external/aladin/client/AladinClientImpl.java
@@ -1,17 +1,15 @@
-package shop.ink3.api.book.book.external.aladin;
+package shop.ink3.api.book.book.external.aladin.client;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import shop.ink3.api.book.book.exception.AladinBookNotFoundException;
