+++ api/src/test/java/shop/ink3/api/book/book/service/BookServiceTest.java
@@ -7,11 +7,10 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import shop.ink3.api.book.author.entity.Author;
 import shop.ink3.api.book.author.repository.AuthorRepository;
-import shop.ink3.api.book.book.dto.BookResponse;
 import shop.ink3.api.book.book.entity.Book;
 import shop.ink3.api.book.book.entity.BookStatus;
 import shop.ink3.api.book.book.exception.BookNotFoundException;
-import shop.ink3.api.book.book.external.aladin.AladinClient;
+import shop.ink3.api.book.book.external.aladin.client.AladinClient;
 import shop.ink3.api.book.book.repository.BookRepository;
 import shop.ink3.api.book.category.entity.Category;
 import shop.ink3.api.book.category.repository.CategoryRepository;
