+++ api/src/test/java/shop/ink3/api/coupon/coupon/service/CouponServiceTest.java
@@ -13,7 +13,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import shop.ink3.api.book.book.entity.Book;
 import shop.ink3.api.book.book.entity.BookStatus;
-import shop.ink3.api.book.book.external.aladin.AladinClientImpl;
+import shop.ink3.api.book.book.external.aladin.client.AladinClientImpl;
 import shop.ink3.api.book.book.repository.BookRepository;
 import shop.ink3.api.book.category.entity.Category;
 import shop.ink3.api.book.category.repository.CategoryRepository;
